feat: Create users through a form
Added react-router and two routes: "/", which renders the student list, and "/new", which renders the student creation form.
The StudentForm should contain all student fields and its respective validations. If any field is invalid, the form can not be submitted.
When submitted, the form will redirect the user to the student list with the brand new student appended into it.